prepare postinstall script
