+++ packages/edge-config/.gitignore
@@ -1,0 +1,1 @@
+stores/*
